DDF-4185 Refactored Splitter Interfaces and Exception class into the API bundle
