+++ catalog/ui/catalog-ui-search/src/main/java/org/codice/ddf/catalog/ui/metacard/ListApplication.java
@@ -47,11 +47,11 @@ import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.StringUtils;
 import org.codice.ddf.attachment.AttachmentInfo;
 import org.codice.ddf.attachment.AttachmentParser;
-import org.codice.ddf.catalog.ui.exceptions.StopSplitterExecutionException;
+import org.codice.ddf.catalog.ui.splitter.StopSplitterExecutionException;
 import org.codice.ddf.catalog.ui.metacard.impl.StorableResourceImpl;
-import org.codice.ddf.catalog.ui.metacard.internal.Splitter;
-import org.codice.ddf.catalog.ui.metacard.internal.SplitterLocator;
-import org.codice.ddf.catalog.ui.metacard.internal.StorableResource;
+import org.codice.ddf.catalog.ui.splitter.Splitter;
+import org.codice.ddf.catalog.ui.splitter.SplitterLocator;
+import org.codice.ddf.catalog.ui.splitter.StorableResource;
 import org.codice.ddf.platform.util.TemporaryFileBackedOutputStream;
 import org.codice.ddf.platform.util.uuidgenerator.UuidGenerator;
 import org.slf4j.Logger;
