+++ ui/packages/catalog-ui-search/src/main/webapp/styles/vars.less
@@ -13,7 +13,7 @@
 //screensizes
 @minimumScreenSize: 20rem; //320px
 @mobileScreenSize: 26.25rem; //420px
-@smallScreenSize: 58.75rem; //940px
+@smallScreenSize: 58.75rem; //940pxF
 @mediumScreenSize: 90rem; // 1440px
 
 //z-indexes
